move get_ids to poplar
