+++ forest/sycamore/common.py
@@ -9,6 +9,7 @@ import glob
 import numpy as np
 import pandas as pd
 
+from forest.poplar.functions.helpers import get_ids
 from forest.sycamore.constants import (EARLIEST_DATE, QUESTION_TYPES_LOOKUP,
                                        ANDROID_NULLABLE_ANSWER_CHANGE_DATE)
 
@@ -880,21 +881,3 @@ def read_aggregate_answers_stream(
     return aggregated_data.loc[
         aggregated_data["answer"] != "NOT_PRESENTED", :
     ]
-
-
-def get_ids(study_folder: str) -> list:
-    """Get subfolders in directory, excluding registry and hidden folders
-
-    Args:
-        study_folder(str): Filepath to the folder containing desired
-            subdirectories
-
-    Returns:
-        List of subdirectories of the study_folder.
-    """
-    list_of_dirs = []
-    for subdir in os.listdir(study_folder):
-        is_folder = os.path.isdir(os.path.join(study_folder, subdir))
-        if (not subdir.startswith(".")) and subdir != "registry" and is_folder:
-            list_of_dirs.append(subdir)
-    return list_of_dirs
