+++ forest/sycamore/tests/test_functions.py
@@ -10,7 +10,6 @@ from forest.sycamore.common import (aggregate_surveys,
                                     aggregate_surveys_no_config,
                                     aggregate_surveys_config,
                                     filename_to_timestamp,
-                                    get_ids,
                                     read_user_answers_stream,
                                     read_aggregate_answers_stream)
 from forest.sycamore.submits import gen_survey_schedule, survey_submits
@@ -264,8 +263,3 @@ def test_file_to_datetime():
     assert filename_to_timestamp(test_str3, "UTC") == expected_timestamp
     test_str4 = "2022-03-14 17_32_56+01_00_1.csv"
     assert filename_to_timestamp(test_str4, "UTC") == expected_timestamp
-
-
-def test_get_ids():
-    users_list = get_ids(SAMPLE_DIR)
-    assert set(users_list) == {"16au2moz", "idr8gqdh"}
